kiet làm xong chức năng coupons và review nhưng giao diện đang fix cứng mai sẽ làm nốt

Đã duyệt

+++ src/main/java/com/vn/ebookstore/service/PaymentDetailService.java
@@ -1,10 +1,11 @@
 package com.vn.ebookstore.service;
 
 import com.vn.ebookstore.model.PaymentDetail;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface PaymentDetailService {
-    PaymentDetail createPayment(Integer orderId, String paymentMethod, Long amount);
+    PaymentDetail createPayment(Integer orderId, String paymentMethod, BigDecimal amount);
     PaymentDetail getPaymentById(Integer id);
     List<PaymentDetail> getPaymentsByOrderId(Integer orderId);
     PaymentDetail updatePaymentStatus(Integer paymentId, String status);
